Session

+++ server/routes/lineItems.js
@@ -9,15 +9,15 @@ const {
 // Will add line item to db of session
 router.post('/', (req, res, next) => {
   const { productId, quantity, cartId } = req.body;
-  const {user, lineItems} = req.session;
+  const { user, lineItems } = req.session;
 
-  if(!user && !lineItems) {
+  if (!user && !lineItems) {
     req.session.lineItems = [req.body];
     res.send(req.session.lineItems);
-  } else if(!user) {
+  } else if (!user) {
     const lineItem = lineItems.find(item => item.productId === productId);
-    if(lineItem) lineItem.quantity += quantity;
-    else lineItems.push(req.body)
+    if (lineItem) lineItem.quantity += quantity;
+    else lineItems.push(req.body);
     res.send(req.session.lineItems);
   } else {
     LineItem.create({ productId, quantity, cartId })
@@ -31,9 +31,9 @@ router.post('/', (req, res, next) => {
 // GET :/api/lineItems/
 // Will get the lineItems saved on the session
 router.get('/', (req, res, next) => {
-  const {lineItems} = req.session;
+  const { lineItems } = req.session;
   if (lineItems) res.send(lineItems);
-})
+});
 
 // GET :/api/lineItems/:cartId
 router.get('/:cartId', (req, res, next) => {
@@ -53,11 +53,17 @@ router.put('/:lineItemId', (req, res, next) => {
     .catch(next);
 });
 
-
 // DELETE :/api/lineItems/:lineItemId
-router.delete('/:lineItemId', (req, res, next) => {
-  const { lineItemId } = req.params;
-  LineItem.destroy({ where: { id: lineItemId } })
-    .then(() => res.sendStatus(204))
-    .catch(next);
+router.delete('/:lineItemId/:productId', (req, res, next) => {
+  const { lineItemId, productId } = req.params;
+  const id = parseInt(productId);
+  let { user, lineItems } = req.session;
+  if (!user) {
+    req.session.lineItems = lineItems.filter(item => item.productId !== id);
+    res.sendStatus(204);
+  } else {
+    LineItem.destroy({ where: { id: lineItemId } })
+      .then(() => res.sendStatus(204))
+      .catch(next);
+  }
 });
